refactor(complete): Pull out completer
